ADDED  Jest.js  configuration and setup files for testing and ts configuration to run tests.
Jest depedencies in package.json e package-lock.json.

upload-writing.tsx now handles just the error state and the UI.

postWriting.ts will be the file with the function responsable to validate and Upload using a HTTP post request

env.ts is a useful file to mock an endpoint while testing.